Implement the whole mod.
Using a dynamic registry. Not sure if they're supposed to not update when datapacks change through the `/datapack enable/disable` commands, but it doesn't seem to be doing that.
Nobody in the fabric discord is answering that question so for now I'm gonna push this and if I get any more info, i'll either fix whatever i'm doing wrong with the dynamic registry, or switch to using a reload listener.